+++ CHANGES.md
@@ -5,6 +5,7 @@
 
 -   Added a function to update the request count of an organisation to match
     its completed tasks.
+-   Check status view now also correctly updates the status
 
 
 ## 1.57 (2025-11-12)